+++ pages/manageFormula.php
@@ -276,10 +276,12 @@ if($_GET['action'] == 'delete' && $_GET['fid']){
 		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> formula '.base64_decode($fid).' is protected.</div>';
 		return;
 	}
-	
+		$meta = mysqli_fetch_array(mysqli_query($conn, "SELECT id FROM formulasMetaData WHERE fid = '$fid'"));
+
 	if(mysqli_query($conn, "DELETE FROM formulas WHERE fid = '$fid'")){
 		mysqli_query($conn, "DELETE FROM formulasMetaData WHERE fid = '$fid'");
-		mysqli_query($conn,"DELETE FROM formulasRevisions WHERE fid = '$fid'");
+		mysqli_query($conn, "DELETE FROM formulasRevisions WHERE fid = '$fid'");
+		mysqli_query($conn, "DELETE FROM formula_history WHERE fid = '".$meta['id']."'");
 		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Formula '.base64_decode($fid).' deleted!</div>';
 	}else{
 		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error</strong> deleting '.base64_decode($fid).' formula!</div>';