+++ pages/manageFormula.php
@@ -252,12 +252,13 @@ if($_POST['action'] == 'addFormula'){
 	$notes = mysqli_real_escape_string($conn, $_POST['notes']);
 	$profile = mysqli_real_escape_string($conn, $_POST['profile']);
 	$catClass = mysqli_real_escape_string($conn, $_POST['catClass']);
+	$finalType = mysqli_real_escape_string($conn, $_POST['finalType']);
 	$fid = base64_encode($name);
 	
 	if(mysqli_num_rows(mysqli_query($conn, "SELECT fid FROM formulasMetaData WHERE fid = '$fid'"))){
 		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error: </strong>'.$name.' already exists!</div>';
 	}else{
-		$q = mysqli_query($conn, "INSERT INTO formulasMetaData (fid, name, notes, profile, catClass) VALUES ('$fid', '$name', '$notes', '$profile', '$catClass')");
+		$q = mysqli_query($conn, "INSERT INTO formulasMetaData (fid, name, notes, profile, catClass, finalType) VALUES ('$fid', '$name', '$notes', '$profile', '$catClass', '$finalType')");
 		if($nID = mysqli_fetch_array(mysqli_query($conn, "SELECT id FROM formulasMetaData WHERE fid = '$fid'"))){
 			echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong><a href="?do=Formula&id='.$nID['id'].'">'.$name.'</a></strong> added!</div>';
 		}else{