+++ db/updates/update_3.6-3.7.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `allergens` ADD `ec` VARCHAR(255) NULL AFTER `cas`;