+++ db/pvault.sql
@@ -6,6 +6,7 @@ CREATE TABLE `allergens` (
   `ing` varchar(255) COLLATE utf8_bin NOT NULL,
   `name` varchar(255) COLLATE utf8_bin NOT NULL,
   `cas` varchar(255) COLLATE utf8_bin DEFAULT NULL,
+  `ec` varchar(255) COLLATE utf8_bin DEFAULT NULL,
   `percentage` DECIMAL(8,4) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;