+++ pages/sellFormula.php
@@ -54,8 +54,8 @@ if($_POST['formula']){
                   </thead>
                   <?php while ($formula = mysqli_fetch_array($formula_q)) {
 					    $ing_q = mysqli_fetch_array(mysqli_query($conn, "SELECT cas,$defCatClass,odor FROM ingredients WHERE name = '".$formula['ingredient']."'"));
-						$conc = number_format($formula['quantity']/$mg['total_mg'] * 100, 3);
-					  	$conc_p = number_format($formula['concentration'] / 100 * $conc, 3);
+						$conc = number_format($formula['quantity']/$mg['total_mg'] * 100, $settings['qStep']);
+					  	$conc_p = number_format($formula['concentration'] / 100 * $conc, $settings['qStep']);
 				  ?>
 					<tr>
                       <td align="center"><?php echo $formula['ingredient']; ?></td>