+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Delete history log when formula is deleted
 Added final type when adding a new formula
 Added purpoce category when adding a new formula
 Fix a bug preventing showing the corrct final type in formula details
+Added EINECS (EC) for ingredient compositions
 
 ### Version 3.6
 Prevent background formula changes when protected mode is enabled