+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added purpoce category when adding a new formula
 Fix a bug preventing showing the corrct final type in formula details
 Added EINECS (EC) for ingredient compositions
 Remove ingredient name from multi dimension lookup, validate only by CAS number
+Enable formula lock/unlock by clicking in lock pad
 
 ### Version 3.6
 Prevent background formula changes when protected mode is enabled